Finished Fisher Blob VFX, Fixed bugs
Hook: Ball now targets blob.x_center, not blob.x_center - 25.

Added Fishing pole for Hook ability

Fixed a bug: Setting player HP to 1 HP (not 1* HP) will result it being read as 0 HP. This has been fixed.

Fixed a bug: After pressing "Reset to Default" on the ruleset screen, changes to either player modifier will modify the other. This lasted until reset. Feature now works as intended.

+++ engine/blobs.py
@@ -124,7 +124,7 @@ class Blob:
         self.image = species_to_image(species)
         self.ability_icon = species_to_ability_icon(species)
         self.stars = species_to_stars(species, stat_overrides) #Gets many values for each blob
-        self.max_hp = 2 * (self.stars['max_hp'] + 3) #Each star adds an additional HP.
+        self.max_hp = int(2 * (self.stars['max_hp'] + 3)) #Each star adds an additional HP.
         self.hp = self.max_hp
         self.top_speed = 10+(1*self.stars['top_speed']) #Each star adds some speed
         self.base_top_speed = self.top_speed #Non-boosted
